requirements.txt
duckdb>=0.9.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
jinja2>=3.1.0
python-multipart>=0.0.6
google-cloud-bigquery>=3.11.0
google-api-python-client>=2.100.0